Gateway Service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,7 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using Contracts;
 using Mapster;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Wolverine;
@@ -56,12 +57,14 @@ public class AuctionsController(AuctionDbContext context, IDbContextOutbox<Aucti
         return auction;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = createAuctionDto.Adapt<Auction>();
-        //TODO: add current user as seller
-        auction.Seller = "TODO: seller";
+
+        
+        auction.Seller = User.Identity?.Name ?? throw new Exception("User has not been Registered");
 
         context.Auctions.Add(auction);
         
@@ -71,6 +74,7 @@ public class AuctionsController(AuctionDbContext context, IDbContextOutbox<Aucti
         return CreatedAtAction(nameof(GetAuction), new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateAuction(string id, UpdateAuctionDto updateAuctionDto)
     {
@@ -89,15 +93,24 @@ public class AuctionsController(AuctionDbContext context, IDbContextOutbox<Aucti
         {
             return BadRequest("Cannot update an auction that has bids");
         }
-        // TODO: Check Seller is the same as the current user
+
+        if (auction.Seller != User.Identity?.Name) return Forbid();
+
         auction.UpdatedAt = DateTime.UtcNow;
-        var updatedAuction = updateAuctionDto.Adapt(auction.Item);
-        await outbox.PublishAsync(updatedAuction.Adapt<AuctionUpdated>());
+
+        updateAuctionDto.Adapt(auction.Item);
+
+        var message = auction.Item.Adapt<AuctionUpdated>();
+        message.Id = auction.Id;
+
+        await outbox.PublishAsync(message);
         await outbox.SaveChangesAndFlushMessagesAsync();
-        
+
         return NoContent();
     }
 
+    
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteAuction(string id)
     {
@@ -111,14 +124,23 @@ public class AuctionsController(AuctionDbContext context, IDbContextOutbox<Aucti
         {
             return BadRequest("Cannot delete an auction that has bids");
         }
-        // TODO: Check Seller is the same as the current user
+
+        if (auction.Seller != User.Identity?.Name) return Forbid();
+        
         context.Auctions.Remove(auction);
+        
         await outbox.PublishAsync(auction.Adapt<AuctionDeleted>());
         await outbox.SaveChangesAndFlushMessagesAsync();
-        
+
         return NoContent();
-        
     }
-    
+
+    [Authorize]
+    [HttpPost("test")]
+    public ActionResult<string> AuthTest()
+    {
+        var name = User.Identity?.Name;
+        return Ok($"{name} has been authenticated");
+    }
     
 }
